+++ src/_data/letterboxd.js
@@ -1,10 +1,10 @@
-require("dotenv").config();
+import "dotenv/config";
 
-const EleventyFetch = require("@11ty/eleventy-fetch");
-const cheerio = require("cheerio");
-const dayjs = require("dayjs");
-const utc = require("dayjs/plugin/utc");
-const relativeTime = require("dayjs/plugin/relativeTime");
+import EleventyFetch from "@11ty/eleventy-fetch";
+import cheerio from "cheerio";
+import dayjs from "dayjs";
+import utc from "dayjs/plugin/utc.js";
+import relativeTime from "dayjs/plugin/relativeTime.js";
 
 dayjs.extend(utc);
 dayjs.extend(relativeTime);
@@ -33,7 +33,7 @@ const fetchRecentMovies = async () => {
 
   $("channel")
     .children("item")
-    .slice(0, 6)
+    .slice(0, 5)
     .each((_, element) => {
       const title = $(element).children("letterboxd\\:filmTitle").text();
 
@@ -65,4 +65,4 @@ const fetchRecentMovies = async () => {
   return recentMovies;
 };
 
-module.exports = fetchRecentMovies;
+export default fetchRecentMovies;