+++ src/_data/navigation.js
@@ -1,24 +1,41 @@
-module.exports = {
+export default {
   top: [
     {
       text: "About",
       url: "/about",
+      icon: "circle-info",
     },
     {
       text: "Posts",
       url: "/posts",
+      icon: "list",
     },
     {
-      text: "Catalogue",
-      url: "/catalogue",
+      text: "Games",
+      url: "/games",
+      icon: "game-controller",
+    },
+    {
+      text: "Watching",
+      url: "/watching",
+      icon: "tv-retro",
+    },
+    {
+      text: "Books",
+      url: "/books",
+      icon: "book",
     },
   ],
   bottom: [
+    { text: "Blogroll", url: "/blogroll" },
     {
       text: "Colophon",
       url: "/colophon",
     },
-    { text: "Blogroll", url: "/blogroll" },
+    {
+      text: "Contact",
+      url: "/contact",
+    },
     {
       text: "/uses",
       url: "/uses",