+++ src/catalogue/catalogue.11tydata.js
@@ -1,6 +1,0 @@
-module.exports = {
-  tags: "catalogue",
-  eleventyComputed: {
-    description: (data) => `My thoughts on ${data.title}.`,
-  },
-};