+++ src/catalogue/games/games.11tydata.js
@@ -1,8 +1,0 @@
-module.exports = {
-  layout: "layouts/catalogue-item",
-  tags: "game",
-  permalink: "catalogue/games/{{ page.fileSlug }}/index.html",
-  eleventyComputed: {
-    subtitle: (data) => `${data.platform}`,
-  },
-};