+++ src/catalogue/podcasts/podcasts.11tydata.js
@@ -1,6 +1,0 @@
-module.exports = {
-  layout: "layouts/catalogue-item",
-  tags: "podcast",
-  permalink: "catalogue/podcasts/{{ page.fileSlug }}/index.html",
-  linkTitle: "Listen to the podcast",
-};