+++ src/content/books/books.11tydata.js
@@ -1,0 +1,5 @@
+export default {
+  layout: "layouts/book",
+  permalink: "books/{{ page.fileSlug }}/index.html",
+  tags: "book",
+};