+++ src/content/games/games.11tydata.js
@@ -1,0 +1,5 @@
+export default {
+  layout: "layouts/game",
+  permalink: "games/{{ page.fileSlug }}/index.html",
+  tags: "game",
+};