+++ src/content/movies/movies.11tydata.js
@@ -1,0 +1,5 @@
+export default {
+  layout: "layouts/movie",
+  permalink: "watching/movies/{{ page.fileSlug }}/index.html",
+  tags: "movie",
+};