+++ src/content/podcasts/podcasts.11tydata.js
@@ -1,0 +1,5 @@
+export default {
+  tags: "podcast",
+  permalink: false,
+  excludeFromSitemap: true,
+};