+++ src/content/shows/shows.11tydata.js
@@ -1,0 +1,5 @@
+export default {
+  layout: "layouts/show",
+  permalink: "watching/shows/{{ page.fileSlug }}/index.html",
+  tags: "tv",
+};